Moved makePromise and oathbreaker to separate files
