+++ lib/Unexpected.js
@@ -7,7 +7,8 @@ var extend = utils.extend;
 var leven = require('leven');
 var cloneError = utils.cloneError;
 var Promise = require('bluebird');
-
+var makePromise = require('./makePromise');
+var oathbreaker = require('./oathbreaker');
 
 var anyType = {
     name: 'any',
@@ -562,26 +563,6 @@ Unexpected.prototype.setErrorMessage = function (err) {
     }
 };
 
-function oathbreaker(value) {
-    if (!value || typeof value.then !== 'function') {
-        return null;
-    }
-
-    if (!value.isRejected) {
-        return value;
-    }
-
-    if (value.isRejected()) {
-        value.caught(function () {
-            // Ignore - already handled
-        });
-
-        throw value.reason();
-    } else if (!value.isFulfilled()) {
-        return value;
-    }
-}
-
 Unexpected.prototype.expect = function expect(subject, testDescriptionString) {
     var that = this;
     if (arguments.length < 2) {
@@ -668,55 +649,7 @@ Unexpected.prototype.expect = function expect(subject, testDescriptionString) {
                 that.fail.apply(that, args);
             });
         };
-        wrappedExpect.promise = function (body) {
-            return new Promise(function (resolve, reject) {
-                var runningTasks = 0;
-                var errors = [];
-
-                function finishWhenDone () {
-                    if (runningTasks === 0) {
-                        if (errors.length > 0) {
-                            reject(errors[0]);
-                        } else {
-                            resolve();
-                        }
-                    }
-                }
-
-                var runner = function (cb) {
-                    runningTasks += 1;
-                    return function () {
-                        runningTasks -= 1;
-                        try {
-                            var result = cb.apply(null, arguments);
-                            var promise = oathbreaker(result);
-                            if (promise) {
-                                runningTasks += 1;
-                                result.then(function () {
-                                    runningTasks -= 1;
-                                    finishWhenDone();
-                                }).caught(function (e) {
-                                    errors.push(e);
-                                    runningTasks -= 1;
-                                    finishWhenDone();
-                                });
-                            }
-                        } catch (e) {
-                            errors.push(e);
-                        } finally {
-                            finishWhenDone();
-                        }
-                    };
-                };
-
-                try {
-                    body(runner);
-                } catch (e) {
-                    errors.push(e);
-                    finishWhenDone();
-                }
-            });
-        };
+        wrappedExpect.promise = makePromise;
 
         wrappedExpect.format = this.format;
         wrappedExpect.it = this.it.bind(this);
